New dashboard update
The following updates contains: New way to store images. It now stores 4 heatmaps in a folder labeled the moments date, every 120 seconds. This will later be used in the dashboard gathering the folders with the 4 images and making a way to browse them. The 4 images is paired together and is a moment together. This will be represented by auto rolling through the 4 images in the moment in the dashboard but also in the future add controllers to view the 4 images in the moment. Also all moments containing 4 images will be designed and based upon how stressed the user was.